+++ packages/wohoox-react/src/createStore.ts
@@ -1,12 +1,12 @@
 import type { ActionsDefine, Options, WohooxPlugin } from 'wohoox'
-import { createStore as createWohooxStore, defaultStoreName } from 'wohoox'
+import { createStore as createWohooxStore, DefaultStoreName } from 'wohoox'
 
 import effectListPlugin from './plugins/effectList'
 
 export function createStore<
   S extends object,
   A extends ActionsDefine<S>,
-  N extends string = typeof defaultStoreName,
+  N extends string = typeof DefaultStoreName,
 >({
   name,
   initState,