+++ packages/wohoox-react/src/hooks/useStore.ts
@@ -1,24 +1,24 @@
 import { useState, useEffect } from 'react'
 
-import { storeMap } from 'wohoox'
+import { getStoreByName, getStoreNames } from 'wohoox'
 
 export function useStore(storeName: string) {
-  const [store, setStore] = useState(() => storeMap.get(storeName))
+  const [store, setStore] = useState(() => getStoreByName(storeName))
 
   if (!store) {
-    const storeKeys = [...storeMap.keys()]
+    const storeKeys = getStoreNames()
 
     if (!storeKeys.length)
       throw new Error(
         'Please check the store has been initialized correctly. \nget store:【' +
           storeName +
-          `】\nexist stores:【'${[...storeMap.keys()].toString()}'】`,
+          `】\nexist stores:【'${getStoreNames().toString()}'】`,
       )
 
     throw new Error(
       'cannot get the correct store。\nget store:【' +
         storeName +
-        `】\nexist stores:【'${[...storeMap.keys()].toString()}'】`,
+        `】\nexist stores:【'${getStoreNames().toString()}'】`,
     )
   }
 
@@ -26,7 +26,7 @@ export function useStore(storeName: string) {
     // For dev environment more，
     // StoreName would not be changed normally.Unless manually changed in dev
     // or set dynamic storeName in production
-    const newStore = storeMap.get(storeName)
+    const newStore = getStoreByName(storeName)
 
     if (store === newStore) return