+++ packages/wohoox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # wohoox
 
+## 2.0.0-rc.5
+
+### Patch Changes
+
+- change the way to get global variables
+
 ## 2.0.0-rc.4
 
 ### Patch Changes